% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtreefit.R
\name{fit_tree}
\alias{fit_tree}
\title{Fit tree to infer branch timings based on observed mutation count and sensitivity}
\usage{
fit_tree(
  tree,
  switch_nodes,
  xcross,
  b_pool_rates = FALSE,
  type = "local",
  niter = 20000,
  suffix = "",
  cores = 3,
  split_nodes = switch_nodes,
  model = "nb_tree"
)
}
\arguments{
\item{tree}{Augmented ape phylo object. A list that should contain agedf - a dataframe with tip.label and age specifying the terminal timepoint for each tip.}

\item{switch_nodes}{Nodes where a change rate occurs}
}
\value{
A list
}
\description{
Fit tree to infer branch timings based on observed mutation count and sensitivity
}
