% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtreefit.R
\name{nbfit_tree_setup_stan_data}
\alias{nbfit_tree_setup_stan_data}
\title{Sets up input data for the stan models.}
\usage{
nbfit_tree_setup_stan_data(
  tree,
  tip_heights,
  rate_switch_nodes,
  xcross = rep(0.95, length(rate_switch_nodes)),
  b_pool_rates = FALSE
)
}
\description{
Sets up input data for the stan models.
}
